pyproject.toml
[tool.poetry]
name = "llm learning"
version = "0.1.0"
description = "LLM learning using GenAI"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = "^2.7.1"
langchain = "^0.3.26"
langchain-core = "^0.3.66"
langchain-openai = "^0.3.24"
openai = "^1.90.0"
langchain-anthropic = "^0.3.15"
langchain-google-genai = "^2.1.5"
langchain-huggingface = "^0.3.0"
transformers = "^4.52.4"
huggingface-hub = "^0.33.0"
python-dotenv = "^1.1.0"
numpy = "^2.3.0"
scikit-learn = "^1.7.0"
streamlit = "^1.46.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
